feat(routeProvider): Add support to catch-all parameters in routes
This allows routeProvider to accept parameters that matches substrings even
when they contain slashes if they are prefixed with an asterisk instead of
a colon.

This PR introduces a new syntax for route params, so we would have:

:param for a single param
*param for catch-all params

For example, routes like edit/color/:color/largecode/*largecode will match
with something like this:
http://appdomain.com/edit/color/brown/largecode/code/with/slashs

It also matches catch-all routes in the middle of the path.

BREAKING CHANGE: Asterisk is not allowed anymore as part of the route
definition